Bewerken en verwijderen van evenement werkt

+++ views/backend/events/events_list.blade.php
@@ -66,7 +66,7 @@
                                     <td>
                                         <div class="btn-group">
                                             <a href="Edit_event.php?id={{$row['id']}}" class="btn btn-default btn-flat"><i class="fa fa-pencil"></i></a>
-                                            <a href="Delete_event_action.php?id={{$row['id']}}" class="btn btn-danger btn-flat"><i class="fa fa-trash"></i></a>
+                                            <a href="Delete_event_action.php?id={{$row['id']}}" onclick="return confirm('Weet je het zeker?')" class="btn btn-danger btn-flat"><i class="fa fa-trash"></i></a>
                                         </div>
                                     </td>
                                 </tr>
